sesuatu

+++ app/Http/Controllers/API/PostController.php
@@ -8,6 +8,7 @@ use App\Models\Like;
 use App\Models\Post;
 use App\Models\Tag;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Storage;
 use Illuminate\Support\Facades\Validator;
 use Illuminate\Validation\ValidationException;
 
@@ -20,7 +21,7 @@ class PostController extends Controller
      */
     public function index()
     {
-        $posts = Post::paginate(10);
+        $posts = Post::with('user:id,name')->with('tags:id,name')->paginate(9);
         $postsData = $posts->items();
         $nextPageUrl = $posts->nextPageUrl();
         $prevPageUrl = $posts->previousPageUrl();
@@ -78,9 +79,10 @@ class PostController extends Controller
                 'message' => 'Post Berhasil Di Buat.',
             ]);
         } catch (ValidationException $e) {
+            $errors = $e->validator->errors()->first();
             return response()->json([
                 'status' => 'gagal',
-                'errors' => $e->errors()
+                'errors' => $errors,
             ], 422);
         } catch (\Exception $e) {
             return response()->json([
@@ -98,7 +100,7 @@ class PostController extends Controller
      */
     public function show(Post $post)
     {
-        $comments = Comment::where('post_id', $post->id)->with('user:id,name')->get();
+        $comments = Comment::where('post_id', $post->id)->with('user:id,name')->with('replies.user:id,name')->get();
         $posts = Post::with('tags:id,name')->with('user:id,name')->find($post->id);
         $post->views++;
         $post->save();
@@ -109,7 +111,7 @@ class PostController extends Controller
             'comment' => $comments
         ]);
     }
-
+    
     /*
      * Update the specified resource in storage.
      *
@@ -119,20 +121,39 @@ class PostController extends Controller
     {
         $validatedData = $request->validate([
             'title' => 'nullable|string|min:3',
-            'content' => 'nullable|string|max:255'
+            'content' => 'nullable|string|max:255',
+            'image' => 'nullable|image'
         ],[
             'title.string' => 'Title harus berupa string.',
             'title.min' => 'Title harus lebih dari 3 karakter atau lebih.',
 
             'content.string' => 'Content harus berupa string.',
-            'content.max' => 'Content terlalu panjang. Maksimal 255 karakter.'
+            'content.max' => 'Content terlalu panjang. Maksimal 255 karakter.',
+            'image.image' => 'Image harus berupa gambar'
         ]);
         try {
             $posts = Post::findOrFail($post);
+            if ($posts->user_id != auth()->user()->id) {
+                return response([
+                    'status' => 'gagal',
+                    'message' => 'Anda tidak bisa meng-update postingan orang lain'
+                ], 403);
+            }
             $posts->title = $validatedData['title'] ?? $posts->title;
             $posts->content = $validatedData['content'] ?? $posts->content;
-            $posts->user_id = $request->user()->id;
-            $posts->save();
+            if($request->hasFile('image')) {
+                $image = $request->file('image');
+                $imageName = $image->getClientOriginalName();
+                $image->storeAs('public/images/posts',$imageName);
+                if ($posts->image !== null) {
+                    Storage::delete('public/images/posts/' . $posts->image);
+                }
+                $validatedData['image'] = $imageName;
+            } else {
+                $validatedData['image'] = $posts->image;
+            }
+            $posts->image = $validatedData['image'] ?? $posts->image;
+            $posts->update();
 
             $tagId = $request->input('tag_id');
             $posts->tags()->sync($tagId);
@@ -158,13 +179,22 @@ class PostController extends Controller
     public function destroy($post)
     {
         $posts = Post::findOrFail($post);
+        if ($posts->user_id != auth()->user()->id) {
+            return response([
+                'status' => 'gagal',
+                'message' => 'Anda tidak bisa menghapus postingan orang lain'
+            ], 403);
+        }
+        if ($posts->image != null) {
+            Storage::delete('public/images/posts/'. $posts->image);
+        }
         $posts->tags()->detach();
         $posts->delete();
 
         return response()->json([
             'status' => 'sukses',
             'message' => 'Post Berhasil Di Hapus.',
-        ],204);
+        ],200);
     }
 
     public function views($postId) {
@@ -172,6 +202,25 @@ class PostController extends Controller
         $view = $post->views;
         return response()->json([
             'viewer' => $view,
+        ]);
+    }
+
+    public function list() {
+        $post = Post::where('user_id', auth()->user()->id)->get();
+        return response()->json([
+            'status' => 'sukses',
+            'message' => 'Menampilkan Post Yang Di Buat',
+            'data' => $post,
+        ]);
+    }
+
+    public function taglist(Tag $id)
+    {
+        $posts = $id->tags()->get();
+        return response()->json([
+            'status' => 'sukses',
+            'message' => 'Menampilkan Post Dengan Tag Tertentu.',
+            'data' => $posts
         ]);
     }
 }
